Refactor turn finalization into dedicated module
