+++ src/agent/runloop/unified/turn/mod.rs
@@ -1,6 +1,8 @@
 mod config_modal;
+mod finalization;
 mod harmony;
 mod session;
+mod utils;
 mod workspace;
 
 pub(crate) use session::run_single_agent_loop_unified;
