+++ src/agent/runloop/unified/turn/session.rs
@@ -22,9 +22,7 @@ use vtcode_core::llm::provider::{self as uni};
 use vtcode_core::tools::registry::{ToolErrorType, ToolExecutionError, classify_error};
 use vtcode_core::ui::slash::{SLASH_COMMANDS, SlashCommandInfo};
 use vtcode_core::ui::theme;
-use vtcode_core::ui::tui::{
-    InlineEvent, InlineEventCallback, InlineHandle, spawn_session, theme_from_styles,
-};
+use vtcode_core::ui::tui::{InlineEvent, InlineEventCallback, spawn_session, theme_from_styles};
 use vtcode_core::utils::ansi::{AnsiRenderer, MessageStyle};
 use vtcode_core::utils::at_pattern::parse_at_patterns;
 use vtcode_core::utils::session_archive::{
@@ -51,7 +49,9 @@ use crate::agent::runloop::unified::ui_interaction::{
 };
 
 use super::config_modal::{MODAL_CLOSE_HINT, load_config_modal_content};
+use super::finalization::finalize_session;
 use super::harmony::strip_harmony_syntax;
+use super::utils::{render_hook_messages, safe_force_redraw};
 use super::workspace::{
     bootstrap_config_files, build_workspace_index, load_workspace_files, refresh_vt_config,
 };
@@ -96,11 +96,9 @@ use crate::agent::runloop::unified::tool_summary::{
     describe_tool_action, humanize_tool_name, render_tool_call_summary_with_status,
 };
 use crate::agent::runloop::unified::workspace_links::{
-    LinkedDirectory, handle_workspace_directory_command, remove_directory_symlink,
+    LinkedDirectory, handle_workspace_directory_command,
 };
-use crate::hooks::lifecycle::{
-    HookMessage, HookMessageLevel, LifecycleHookEngine, SessionEndReason, SessionStartTrigger,
-};
+use crate::hooks::lifecycle::{LifecycleHookEngine, SessionEndReason, SessionStartTrigger};
 use crate::ide_context::IdeContextBridge;
 
 enum TurnLoopResult {
@@ -2607,93 +2605,18 @@ pub(crate) async fn run_single_agent_loop_unified(
             }
         }
 
-        let transcript_lines = transcript::snapshot();
-        if let Some(archive) = session_archive.take() {
-            let distinct_tools = session_stats.sorted_tools();
-            let total_messages = conversation_history.len();
-            let session_messages: Vec<SessionMessage> = conversation_history
-                .iter()
-                .map(SessionMessage::from)
-                .collect();
-            match archive.finalize(
-                transcript_lines,
-                total_messages,
-                distinct_tools,
-                session_messages,
-            ) {
-                Ok(path) => {
-                    if let Some(hooks) = &lifecycle_hooks {
-                        hooks.update_transcript_path(Some(path.clone())).await;
-                    }
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!("Session saved to {}", path.display()),
-                    )?;
-                    renderer.line_if_not_empty(MessageStyle::Output)?;
-                }
-                Err(err) => {
-                    renderer.line(
-                        MessageStyle::Error,
-                        &format!("Failed to save session: {}", err),
-                    )?;
-                    renderer.line_if_not_empty(MessageStyle::Output)?;
-                }
-            }
-        }
-
-        for linked in linked_directories {
-            if let Err(err) = remove_directory_symlink(&linked.link_path).await {
-                eprintln!(
-                    "Warning: failed to remove linked directory {}: {}",
-                    linked.link_path.display(),
-                    err
-                );
-            }
-        }
-
-        if let Some(hooks) = &lifecycle_hooks {
-            match hooks.run_session_end(session_end_reason).await {
-                Ok(messages) => {
-                    render_hook_messages(&mut renderer, &messages)?;
-                }
-                Err(err) => {
-                    renderer.line(
-                        MessageStyle::Error,
-                        &format!("Failed to run session end hooks: {}", err),
-                    )?;
-                }
-            }
-        }
-
-        // Shutdown MCP client properly before TUI shutdown using async manager
-        if let Some(mcp_manager) = &async_mcp_manager
-            && let Err(e) = mcp_manager.shutdown().await
-        {
-            let error_msg = e.to_string();
-            if error_msg.contains("EPIPE")
-                || error_msg.contains("Broken pipe")
-                || error_msg.contains("write EPIPE")
-            {
-                eprintln!(
-                    "Info: MCP client shutdown encountered pipe errors (normal): {}",
-                    e
-                );
-            } else {
-                eprintln!("Warning: Failed to shutdown MCP client cleanly: {}", e);
-            }
-        }
-
-        handle.shutdown();
-
-        // Clear the inline handle from the message queue system
-        transcript::clear_inline_handle();
-
-        // Clean up TUI mode environment variable
-        // SAFETY: We're removing the variable we set at the start of the session.
-        // No other threads should be accessing this variable at this point.
-        unsafe {
-            std::env::remove_var("VTCODE_TUI_MODE");
-        }
+        finalize_session(
+            &mut renderer,
+            lifecycle_hooks.as_ref(),
+            session_end_reason,
+            &mut session_archive,
+            &session_stats,
+            &conversation_history,
+            linked_directories,
+            async_mcp_manager.as_deref(),
+            &handle,
+        )
+        .await?;
 
         // If the session ended with NewSession, restart the loop with fresh config
         if matches!(session_end_reason, SessionEndReason::NewSession) {
@@ -2707,33 +2630,6 @@ pub(crate) async fn run_single_agent_loop_unified(
         }
 
         break;
-    }
-
-    Ok(())
-}
-
-fn safe_force_redraw(handle: &InlineHandle, last_forced_redraw: &mut Instant) {
-    // Rate limit force_redraw calls to prevent TUI corruption
-    if last_forced_redraw.elapsed() > std::time::Duration::from_millis(100) {
-        handle.force_redraw();
-        *last_forced_redraw = Instant::now();
-    }
-}
-
-fn render_hook_messages(renderer: &mut AnsiRenderer, messages: &[HookMessage]) -> Result<()> {
-    for message in messages {
-        let text = message.text.trim();
-        if text.is_empty() {
-            continue;
-        }
-
-        let style = match message.level {
-            HookMessageLevel::Info => MessageStyle::Info,
-            HookMessageLevel::Warning => MessageStyle::Info,
-            HookMessageLevel::Error => MessageStyle::Error,
-        };
-
-        renderer.line(style, text)?;
     }
 
     Ok(())
